+++ components/forum/components/PostCategoryPermissions/FreePostCategoryPermissions/FreeSharePostCategoryToWeb.tsx
@@ -1,4 +1,4 @@
-import { SharePostCategoryToWeb } from '../common/SharePostCategoryToWeb';
+import { SharePostCategoryToWeb } from '../components/SharePostCategoryToWeb';
 
 export function FreeSharePostCategoryToWeb() {
   return <SharePostCategoryToWeb disabled isChecked />;
