+++ components/forum/components/PostCategoryPermissions/shared.ts
